Port website to React Static
The website was originally created with Create React App. This meant
the bundle needed to be downloaded in its entirity and processed before
the page rendered. This resulted in optimizatiion and SEO issues.
Image optimzations further required an npm eject. Hence, the decision
was made to painstakingly port the entire project to React Static.
React Static is a newer project that is maintained by a smaller group of
people. As a result, documentation and support are less prevalant than
with Create React App. That said, the working static file is `blazing`
fast and the freedom to modify webpack configurations existed out of the
box. Gatsby was considered, but the reliance on GraphQL did not fit well
with this project. I needed the ability to make individual image queries
without a CMS system in place.